# apps/web/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "LoginRequest": {
        "additionalProperties": false,
        "description": "Credentials accepted by the local identity-provider boundary.",
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MfaCodeRequest": {
        "additionalProperties": false,
        "description": "A current TOTP or one complete recovery credential.",
        "properties": {
          "code": {
            "maxLength": 80,
            "minLength": 6,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "MfaCodeRequest",
        "type": "object"
      },
      "MfaDeleteRequest": {
        "additionalProperties": false,
        "description": "Exactly one current factor proof accepted before disabling MFA.",
        "properties": {
          "code": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "recovery": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 6,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovery"
          }
        },
        "title": "MfaDeleteRequest",
        "type": "object"
      },
      "MfaEnrollRequest": {
        "additionalProperties": false,
        "description": "Current password required before generating standing MFA credentials.",
        "properties": {
          "password": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "MfaEnrollRequest",
        "type": "object"
      },
      "MfaEnrollmentResponse": {
        "description": "One-time enrollment material that is never persisted in plaintext.",
        "properties": {
          "otpauth_uri": {
            "title": "Otpauth Uri",
            "type": "string"
          },
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "title": "Recovery Codes",
            "type": "array"
          },
          "secret": {
            "title": "Secret",
            "type": "string"
          }
        },
        "required": [
          "secret",
          "otpauth_uri",
          "recovery_codes"
        ],
        "title": "MfaEnrollmentResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "XLR8 FLO API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate credentials and rotate any already authenticated session.",
        "operationId": "login_api_v1_auth_login_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Login",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Immediately revoke the current server-side session.",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Logout",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/mfa": {
      "delete": {
        "description": "Disable a non-mandatory factor after a current proof and rotate the session.",
        "operationId": "disable_mfa_api_v1_auth_mfa_delete",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Disable Mfa",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/mfa/confirm": {
      "post": {
        "description": "Activate a pending factor and rotate the authenticated session identifier.",
        "operationId": "confirm_mfa_api_v1_auth_mfa_confirm_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Confirm Mfa",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/mfa/enroll": {
      "post": {
        "description": "Generate one inactive TOTP secret after current-password re-authentication.",
        "operationId": "enroll_mfa_api_v1_auth_mfa_enroll_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaEnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MfaEnrollmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Enroll Mfa",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "description": "Complete login or step-up, refresh recent auth, and rotate the session.",
        "operationId": "verify_mfa_api_v1_auth_mfa_verify_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MfaCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Verify Mfa",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/reset": {
      "post": {
        "description": "Replace one credential after atomically consuming its reset token.",
        "operationId": "reset_password_api_v1_auth_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetCompletion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Reset Password",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/reset-request": {
      "post": {
        "description": "Accept every syntactically valid account locator without revealing existence.",
        "operationId": "request_password_reset_api_v1_auth_reset_request_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-FLO-Origin-Secret",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Flo-Origin-Secret"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Client error expressed as RFC 9457 problem details."
          },
          "5XX": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Server error expressed as RFC 9457 problem details."
          }
        },
        "summary": "Request Password Reset",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}